services/api/pyproject.toml
[project]
name = "api-service"
version = "0.1.0"
description = "FastAPI-based API"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
  "fastapi",
  "uvicorn",
  "util-math @ file://{root}/../../packages/util_math",
  "data-io @ file://{root}/../../packages/data_io",
  "python-dotenv",
  "boto3",
  "sqlalchemy>=2.0",
  "psycopg[binary]",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.metadata]
allow-direct-references = true

[tool.hatch.build.targets.wheel]
packages = ["api"]
